backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-filter==23.3
markdown==3.5.1
pytz==2023.3
drf-yasg==1.21.7
